+++ crates/ruma-events/src/key/verification/ready.rs
@@ -46,7 +46,7 @@ impl ToDeviceKeyVerificationReadyEventContent {
 /// Response to a previously sent `m.key.verification.request` message.
 #[derive(Clone, Debug, Deserialize, Serialize, EventContent)]
 #[cfg_attr(not(feature = "unstable-exhaustive-types"), non_exhaustive)]
-#[ruma_event(type = "m.key.verification.ready", kind = Message)]
+#[ruma_event(type = "m.key.verification.ready", kind = MessageLike)]
 pub struct KeyVerificationReadyEventContent {
     /// The device ID which is initiating the request.
     pub from_device: Box<DeviceId>,