+++ crates/ruma-events/src/reaction.rs
@@ -9,7 +9,7 @@ use serde::{Deserialize, Serialize};
 /// A reaction to another event.
 #[derive(Clone, Debug, Deserialize, Serialize, EventContent)]
 #[cfg_attr(not(feature = "unstable-exhaustive-types"), non_exhaustive)]
-#[ruma_event(type = "m.reaction", kind = Message)]
+#[ruma_event(type = "m.reaction", kind = MessageLike)]
 pub struct ReactionEventContent {
     /// Information about the related event.
     #[serde(rename = "m.relates_to")]